scraping_agent/requirements.txt
browser-use>=0.12.0
playwright>=1.40.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
